Add custom Material Design theme
Adapted from https://github.com/FezVrasta/bootstrap-material-design;
I've built the theme from source -- written in LESS -- and customized a
few variables in the base theme. I've also overridden a few classes in
the meantime to make them work for this site.

It may make sense to include the source and build steps for this theme
in the repository... that's a point that can be discussed with the team.