Layout temporary pages for main pages
About, Community, and the FAQ page have their basic layout
complete. Docs is still a work-in-progress.

The FAQ page includes the ability to expand and collapse the QAs in an
accordion-like fashion. We're going to eventually pull these QAs from a
seperate data directory with a separate file for each QA that contains
the appropriate markup in markdown.