Basic blocking of layout (#355)
* Remove jekyll-theme-leap-day

* Add custom Material Design theme

Adapted from https://github.com/FezVrasta/bootstrap-material-design;
I've built the theme from source -- written in LESS -- and customized a
few variables in the base theme. I've also overridden a few classes in
the meantime to make them work for this site.

It may make sense to include the source and build steps for this theme
in the repository... that's a point that can be discussed with the team.

* Create base and default layouts

Standard base and default layotus that can then be adapted with
page-specific layouts when necessary. Takes care of all the markup
surrounding the content.

* Add navigation bar

Navigation bar now includes links to the homepage and all subpages:
About, Docs, Community, and FAQ.

The search bar still needs to be added to the navigation bar and it must
adhere to Material Design principles. This requires a little tweaking
on the underlying Material Design theme.

* Layout temporary pages for main pages

About, Community, and the FAQ page have their basic layout
complete. Docs is still a work-in-progress.

The FAQ page includes the ability to expand and collapse the QAs in an
accordion-like fashion. We're going to eventually pull these QAs from a
seperate data directory with a separate file for each QA that contains
the appropriate markup in markdown.

* Prefix CSS files with site's base URL

* Add Forseti logo to navbar

* Link to Forseti favicon

* Fix file indentation

* Layout docs page

Finishing blocking the layout for the docs page.

Added _includes/module_inventory_content.md as a temporary file to test
inclusion of what actual content will look like. The left side bar and
right table of contents (toc) use dummy content.

I will cleanup the inline styles. These are temporary and purely meant
for the purpose of blocking out the layout.

* Layout homepage

Basic blocking of layout on the homepage.

Further input required from the team on exactly what copy/images/layout
is necessary for this page.

* Add third-party code for bootstrap material design

Also, exclude all repos in third_party/ from Jekyll site generation.

* Fix indentation to Google HTML/CSS style guide

Adjust indentation in all applicable files to be 2 spaces per tab, as
prescribed by https://google.github.io/styleguide/htmlcssguide.html.

* Fix CSS selectors and property ordering

Bring CSS selectors and property ordering inline with Google's HTML/CSS
style guide.

* Include protocol on link.href's

Per Google's HTML/CSS style guide.

* Clean unnecessary files from third_party/

Remove all third_party files from bootstrap-material-design that are not
absolutely required to be in the forseti-security repository.

If you want to rebuild a custom version of bootstrap-material-design,
you must do it out-of-band, and then copy the distribution
files (generated in dist/) to
third_party/bootstrap-material-design/dist/ and css/.

* Fix indentation in navbar file