Add a new utility for recording and replaying API requests. (#808)
* Add a new utility for recording and replaying API requests.

This change migrates functionality from gcp_api2/_base_client into
gcp_api.

The record and replay functionality is extended to use a global
dictionary to store the responses so it works across all API
repositories. This allows a developer to record all output from a live
environment and then use the data in testing local on their system.

The environment variables FORSETI_RECORD_FILE and FORSETI_REPLAY_FILE
are introduced to enable the recording to or replaying from a file.