Build IAM policy bindings in the scanner proper
Resources can inherit policy bindings from ancestors in the resource
manager tree. For example: a GCS bucket inherits a 'objectViewer' role
from a project or folder (up in the tree).

So far the IAM rules engine only checks the set of bindings directly
attached to a resource (direct bindings set (DBS)). We need to add
relevant bindings inherited from ancestors to DBS so that these are also
checked for violations.

The only place where we can do this (since we have a full list of
resources and policies) is the scanner proper. That's why I am moving
the construction of IAM policy bindings from the rule engine to the
scanner.

Next I will introduce functionality to find the relevant bindings a
resource inherits from its ancestors (inherited bindings set
(IBS)).

The IAM rules engine will then check both DBS and IBS for violations.