Fix the construction of IAM policy 'full_name'. (#1185)
* Ensure 'full_name' has an even number of segments

The construction of a 'full_name' for 'iam_policy' resources (and
possibly others) is broken since 3 segments are appended to the
'full_name' of the parent resulting e.g. in

    organization/111222333/iam_policy/organization/111222333/

This breaks other code (e.g. get_resources_from_full_name()) assuming
that a full_name consists of pairs of segments i.e.

    organization/111222333/iam_policy/111222333/

Fixes issue #1174.

* Better fix that avoids name conflicts

* ':' is a better separator

':' is an invalid character in project name, bucket name and other GCP
resources.