GCS bucket IAM scanner (#1168)
* IAM scanner now fetches bucket IAM policies from model

* fix typo

* make bucket a valid resource type

* allow for bucket level IAM rules

* ignore unknown policy member types (issue #1162)

* fix pylint warnings

* Only use parent's name for IAM policy full_name

Fixes issue #1174

* Warn about uknown policy member types.

issue #1162

* Give policy parent resources proper full names.

* Build IAM policy bindings in the scanner proper

Resources can inherit policy bindings from ancestors in the resource
manager tree. For example: a GCS bucket inherits a 'objectViewer' role
from a project or folder (up in the tree).

So far the IAM rules engine only checks the set of bindings directly
attached to a resource (direct bindings set (DBS)). We need to add
relevant bindings inherited from ancestors to DBS so that these are also
checked for violations.

The only place where we can do this (since we have a full list of
resources and policies) is the scanner proper. That's why I am moving
the construction of IAM policy bindings from the rule engine to the
scanner.

Next I will introduce functionality to find the relevant bindings a
resource inherits from its ancestors (inherited bindings set
(IBS)).

The IAM rules engine will then check both DBS and IBS for violations.

* Fix unit tests

* Add function that finds bucket relevant bindings

_add_bucket_ancestor_bindings() looks at policy bindings of bucket
ancestors and copies those that are storage relevant to the bucket's
list (of bindings).

This way *all* the bindings that apply to a bucket will be checked for
violations.

* more concise documentation

* Turn on _add_bucket_ancestor_bindings() function

* Ensure 'full_name' has an even number of segments

The construction of a 'full_name' for 'iam_policy' resources (and
possibly others) is broken since 3 segments are appended to the
'full_name' of the parent resulting e.g. in

    organization/111222333/iam_policy/organization/111222333/

This breaks other code (e.g. get_resources_from_full_name()) assuming
that a full_name consists of pairs of segments i.e.

    organization/111222333/iam_policy/111222333/

Fixes issue #1174.

* Better fix that avoids name conflicts

* ':' is a better separator

':' is an invalid character in project name, bucket name and other GCP
resources.

* Add test with "wrong" ancestors

* Test with resource tree and 2 ancestor lines.

* more realistic 'full_path' for policies.

* better naming / portability of test data

* add 2 tests with policy attached to a bucket

* Remove unneeded test data.

* Add IamPolicyScanner::_retrieve() test for buckets

* Add a IamPolicyBinding::merge() method

* For IAM bindings with same role name: merge members

* Rewrite _add_bucket_ancestor_bindings() for readability

* fix pylint warnings

* fix typo

* add documentation

* Rename: merge -> merge_members

* Resolve reviewer comments re. comprehensions

* Fix pylint warning