Fix build script for Travis CI
Travis has been erroneously passing builds that should fail because the docs don't build due
to broken links.

However, the script was not returning a non-zero exit code in the case that the jekyll build
failed; it was returning a zero exit code because the HTML proofer found no nothing wrong
with an empty site.

We fix this error by trapping any non-zero exit codes from intermediate steps, and returning
that exit code to the calling process. Therefore, Travis will be properly notified when the
Jekyll build fails and when the HTML proofer fails.