Add jmespath scanner for KE (#1991)
* Refactor most KE functionality into a base class

Most of this can be reused for subsequent scanner implementation.  Also
add a hook for scanner subclasses to extend the violation_data in a
customizable way.

* When constructing a KeCluster, keep the API's dict around

The jmespath scanner will enable users to write rules that extract
arbitrary fields from a KE cluster's API representation.  That means the
object needs to have that dict to match against.

* Add KE Cluster jmespath scanner

This scanner allows the user to write rules that extract arbitrary values
from a KE cluster's JSON representation, and match them against a given
list of values.  The values can be treated as a whitelist or a blacklist.

Some sample ideas are included in rules/ke_jmespath_rules.yaml

* Add myself to contributors list

* Update config file template

* Update copyright attribution

Adding ClearDATA to AUTHORS, Joe Ceresini to CONTRIBUTORS, and cleaning up
the in-source individual attributions.

* Move KeBaseScanner into its own file

* Add new scanner to forseti_conf_server.yaml.sample

* Remove "jmespath" from names

* Remove advaced sample rule from ke_rules.yaml file

Including this example breaks the forseti installer.  The installer uses
python's str.format to fill in strings of the form "{variable}".  The
brances in this example make python want a template parameter that it
doesn't have.

* Add KE Scanner info to scanner_requirements_map

This also update the KE Version Scanner info to point to the updated name
for the rules file.

* Fix legacyABAC example rule to match description

* Make KE scanner violation message more useful

* Fix improved log messages and unittests

* Build the as_dict property as-required from _json

* Fix typo

* Pluralize resource_rule to match class name

* Additional example rules, and a pointer to the test data

Currently the example rules can't include any "{...}" expressions since
the forseti installer uses str.format to fill in the template.

* Add min version to jmespath requirement

* Use ke_scanner_rules.yaml for the new KE scanner

This avoids breaking existing installations with ke_rules.yaml for
the KE version scanner.

* Create KeBaseScanner subclasses via a factory function

* Fix pylint violations

* Include resource_name in KeScanner violations

* Make the mode parameter mandatory

Users' rules will be easier to understand if they must always include the
mode.  Also update tests to match test data.

* Make KE scanner classes explicit classes