Speed up import of cloudasset data into temporary sql table.
* Strip unused data before writing to the database
* Switch from SQLAlchemy ORM to Core style bulk CAI inserts
* Ensure the maximum number of rows are written on each bulk insert

This is one of a series of changes to reduce the time to complete
a forseti inventory snapshot for large organizations.

This change reduces the time to import data into a cloudsql database
from a test system by about 50%, from about 312 seconds to about 144
seconds for around 350 megabytes of raw data.