Stream CloudAsset data straight to sqlite database.
* Change cloudasset implementation to stream data from GCS to the local
sqlite database using OS pipes instead of downloading the full file to a
temporary location and then reading it from there.

This change reduces storage requirements on Forseti Server and allows
data to be processed only once instead of writing and rereading each row
of data from the CloudAsset export.